Use glow for enchanted arrows (feature #5122)
